fix(safety): preserve fog and bound loaders
Preserve configured night fog colours while hardening loader and spawn data
against malformed fixed-array indices and counts.

- Keep configured RGB fog and transparency values in night mode
- Bound animation, resource, script, spawn, and pack references
- Preserve legacy decimal parsing for health and kill distance fields
- Test: all four CTest trees and deployed GL/software night smoke tests pass

+++ Hunt/Core/GameTypes.h
@@ -639,7 +639,7 @@ struct TWeapInfo
   unsigned char crossRed, crossGreen, crossBlue;
   WORD crossColour565, crossColour555;
 
-  int recoil;
+  float recoil;
 
 };
